Revert broken Add support for URL fragment parameters
This is a revert of the code changes in commit
f796b05e42cfac7044cca9603e59f258605228f3 as it served no functional
purpose.

Fragments were already respected for setting parameters, via a different
function. Thus it is unclear what that commit tried to fix. It also
complicated things by mixing the document location with the window
location.

The comment changes are useful, though, so those are kept.